cmd/bsky-webhook: make dial and login errors non-fatal
Rather than fail the program when a dial or login to Bluesky fails, report the
error back to the server loop and let it retry. We might want to eventually
give up if it fails for "long enough", but for now this will help work around
transient service errors, downtime, etc.